+++ resources/views/layouts/master.blade.php
@@ -7,9 +7,9 @@
 
         @if(null != Cookie::get('uid'))
             <nav class="row top-bar">
-                <a href="/profile" class="profile-pic"><img src="//graph.facebook.com/{{ $user->fbid }}/picture?width=80&height=80"></a>
-                <a href="/profile" class="bold">{{ trans('master.logged_as', ['name' => $user->first_name]) }}</a>
-                <a href="/profile">{{ trans('master.goto') }}</a>
+                <a href="/profile/records" class="profile-pic"><img src="//graph.facebook.com/{{ $user->fbid }}/picture?width=80&height=80"></a>
+                <a href="/profile/records" class="bold">{{ trans('master.logged_as', ['name' => $user->first_name]) }}</a>
+                <a href="/profile/records">{{ trans('master.goto') }}</a>
                 <a href="/profile/logout">{{ trans('master.logout') }}</a>
                 <a href="/" class="logo"><img src="/img/logo-red-80.png" alt=""></a>
             </nav>