+++ resources/views/partials/footer.blade.php
@@ -15,7 +15,9 @@
     <div class="col-4 area-2">
     	<span class="light-grey-text big">{{ trans('master.menu') }}</span>
     	<ul>
+            <a href="/"><li class="grey-text">{{ trans('page.home') }}</li></a>
             <a href="/community"><li class="grey-text">{{ trans('page.community') }}</li></a>
+            <a href="/profile/records"><li class="grey-text">{{ trans('page.profile') }}</li></a>
     		<li class="almost-black-text">{{ trans('master.faq') }}</li>
             <li class="almost-black-text">{{ trans('master.about') }}</li>
     	</ul>